Create 622. Designing a circular queue
Design your circular queue implementation. A circular queue is a linear data structure whose operation performance is based on the FIFO (First In First Out) principle and the tail of the team is connected after the head of the team to form a loop. It is also called a "ring buffer".
One of the benefits of circular queues is that we can use the space that this queue has previously used. In an ordinary queue, once a queue is full, we cannot insert the next element, even if there is still space in front of the queue. But with circular queues, we can use this space to store new values.