Create 724. Finding the Center Index of an Array
Given an array nums of integer type, write a method that returns the "center index" of the array.

We define the array center index like this: the sum of all elements on the left side of the array center index is equal to the sum of all elements on the right side.
If the array does not have a central index, then we should return -1. If the array has multiple center indices, then we should return the one closest to the left.